pyproject.toml
[tool.poetry]
name = "python_ocr"
version = "1.0.0"
description = "Handwritten OCR service based on TrOCR + FastAPI + FTP"
authors = ["Daniel Nuño <dnuno@cetys.edu.mx>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.0" 
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
transformers = "^4.30.0"
torch = "^2.0.1"
numpy = "^1.24.3"
pillow = "^9.5.0"
opencv-python = "^4.7.0"
scipy = "^1.10.1"
matplotlib = "^3.7.1"
supabase = "^1.0.3"
pyftpdlib = "^1.5.8"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
email-validator = "^2.0.0"
resend = "^0.5.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"